requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.10.0
attrs==25.4.0
audioop-lts==0.2.2
beautifulsoup4==4.13.4
blinker==1.9.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
chatbot==1.5.2b0
click==8.2.1
colorama==0.4.6
comtypes==1.4.11
cryptography==46.0.5
curl_cffi==0.13.0
distlib==0.4.0
distro==1.9.0
Django==5.2.7
django-cors-headers==4.9.0
django-rest-framework==0.1.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
filelock==3.20.0
Flask==3.1.2
frozendict==2.4.7
frozenlist==1.8.0
google-ai-generativelanguage==0.6.15
google-api-core==2.30.0
google-api-python-client==2.190.0
google-auth==2.49.0.dev0
google-auth-httplib2==0.3.0
google-genai==1.64.0
google-generativeai==0.8.6
googleapis-common-protos==1.72.0
grpcio==1.78.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
MouseInfo==0.1.3
multidict==6.7.1
multitasking==0.0.12
mysql-connector-python==9.4.0
mysqlclient==2.2.7
numpy==2.4.2
packaging==25.0
pandas==3.0.0
peewee==3.19.0
pefile==2023.2.7
pillow==11.3.0
pipenv==2025.0.4
platformdirs==4.5.0
propcache==0.4.1
proto-plus==1.27.1
protobuf==5.29.6
pyasn1==0.6.2
pyasn1_modules==0.4.2
PyAudio==0.2.14
PyAutoGUI==0.9.54
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
PyGetWindow==0.0.9
pyinstaller==6.15.0
pyinstaller-hooks-contrib==2025.8
pyjokes==0.8.3
PyJWT==2.10.1
PyMsgBox==1.0.9
pyparsing==3.3.2
pyperclip==1.9.0
pypiwin32==223
PyRect==0.2.0
PyScreeze==1.0.1
python-dateutil==2.9.0.post0
pyttsx3==2.99
pytweening==1.2.0
pytz==2025.2
pywhatkit==5.4
pywin32==311
pywin32-ctypes==0.2.3
requests==2.32.5
sambanova==1.4.1
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SpeechRecognition==3.14.3
sqlparse==0.5.3
standard-aifc==3.13.0
standard-chunk==3.13.0
ta==0.11.0
tenacity==9.1.4
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
virtualenv==20.35.3
virtualenvwrapper-win==1.2.7
websockets==15.0.1
Werkzeug==3.1.3
wikipedia==1.4.0
yarl==1.22.0
yfinance==1.1.0
